backend/requirements.txt
langchain==1.2.13
langgraph
langchain-openai
fastapi
uvicorn[standard]
sse-starlette
python-dotenv
pydantic-settings
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
psycopg[binary,pool]
python-jose[cryptography]
bcrypt
langgraph-checkpoint-postgres
sqlalchemy
psycopg2-binary
alembic
